distributed_string_processor/grpc_stub/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   starlette
    #   watchfiles
click==8.1.8
    # via uvicorn
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.12
    # via -r requirements.in
grpcio==1.58.3
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via anyio
jinja2==3.1.6
    # via -r requirements.in
markupsafe==3.0.2
    # via jinja2
protobuf==3.20.3
    # via -r requirements.in
pydantic==2.11.4
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.1.0
    # via uvicorn
python-multipart==0.0.5
    # via -r requirements.in
pyyaml==6.0.2
    # via uvicorn
six==1.17.0
    # via python-multipart
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
typing-extensions==4.13.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.0
    # via pydantic
uvicorn[standard]==0.34.2
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn
